Added Propeller Mushroom powerup.
Unpacked the "LargeMario" and "LargeLuigi" prefab inside the "PlayerMario" and "LuigiPlayer" prefab to add in the propeller helment model. I'm not sure if there was a way to add it to their big models directly I'm still learning how to use Unity.

The Propeller mushroom only spawns if custom powerups are enabled but the spin ability hasn't been re-added. Will do that tomorrow as it's currently late.