ScrollCamera RenderToggle
Created a Script so that when invisible objects are in frame on either the Left or Right Hand sides of the stage, then the Scroll Camera will render.

Otherwise, the Camera will be disabled, saving on performance (in theory, have yet to test outside of the Unity Editor).

Furthermore, this is also exclusively on Grasslands at the moment, but the setup is simple enough to implemented on all looping stages.