Disable Updating of Sprite Renderer & Volume w/ Distance on Dead Enemies and Collected Coins
When objects that use LegacyAnimateSprite Renderer are collected or killed (like coins for example), they still update the Sprite Renderer despite not needing too. This fixes that.

In addition, also disabled VolumeWithDistance for dead enemies or collected coins, since there would be no point in doing so as their sound effects couldn't be played in that state.