Mostly Accurate Spin Block / Spinner
Chances are some of the values are wrong, because I just eyeballed it from a low res screenshot. Also, the model is inaccurate (besides the normals), because the spheremapped bulb on the bottom should rotate with the rest of the model. I hacked around the messy normals by setting it to calculate them on import, instead of importing the original normals. Works for now, but I would really like to know how the spin block is animated, so I can swap out the model completely.